Refactor RocksDB backend to split the module
